+++ src/include/duckdb/common/weak_ptr_ipp.hpp
